backend/requirements.txt
flask
flask-cors
transformers
torch
pillow
pytesseract
requests
